Vectorising the code
Using Armadillo vectors for the particle weights, and updating some of the examples to use Armadillo vectors. A MCMC step has been added to the radiata pine regression example and a likelihood annealing approach for this example has also been added (will be used later to test path sampling integral).