Passing algorithm parameters into moveset and adaptation functions
The algorithm parameters were previously kept in the class used for adaptation but now they are passed by reference to the adaptation and moveset functions. This is simpler and easier to use, especially in the case that there is no adaptation. I think it'd be nice to avoid having to declare a type for the parameters because many examples don't use them (or they are simple enough to just use as global variables).