Fully adapted (auxiliary) particle filter
Adding the fully adapted auxiliary particle filter implementation.

It would be helpful to generalise this in the future (allowing user defined means and covariances for the initial, transition and measurement densities).

It is probably also preferable to make the transition matrix a parameter which can be passed to the move function rather than a global variable.

This example still needs to be documented properly (i.e. with an .Rd file).